separate create_protocol and reader unit tests
